+++ pkg/grpc/actions/canvases/common_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/superplanehq/superplane/pkg/models"
 	"github.com/superplanehq/superplane/test/support"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/protobuf/types/known/structpb"
 )
 
 func setupLiveCanvasStaging(t *testing.T) (*support.ResourceRegistry, context.Context, *models.Canvas, *models.CanvasVersion) {
@@ -27,15 +26,6 @@ func setupLiveCanvasStaging(t *testing.T) (*support.ResourceRegistry, context.Co
 	require.NoError(t, err)
 
 	return r, ctx, canvas, liveVersion
-}
-
-func structFromAnyMap(t *testing.T, value map[string]any) *structpb.Struct {
-	t.Helper()
-
-	result, err := structpb.NewStruct(value)
-	require.NoError(t, err)
-
-	return result
 }
 
 func TestMapCanvasNameUniqueConstraintError(t *testing.T) {
