+++ pkg/grpc/actions/canvases/version_serialization.go
@@ -12,14 +12,6 @@ import (
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
-func canvasMetadataFromCanvas(canvas *models.Canvas) (name, description string) {
-	if canvas == nil {
-		return "", ""
-	}
-
-	return canvas.Name, canvas.Description
-}
-
 func SerializeCanvasVersion(version *models.CanvasVersion, organizationID string, ownersByID map[string]*models.User) (*pb.CanvasVersion, error) {
 	var author *pb.UserRef
 	if version.OwnerID != nil {
